+++ Projekt 2 - github/include/graphs/graph.hpp
@@ -1,17 +1,16 @@
 #ifndef GRAPH_HPP_
 #define GRAPH_HPP_
+
 #include <vector>
-template <typename T>
+
 class Graph
 {
-  public:
-    virtual void add_vertex(T vertex) = 0;//dodanie wierzcholka
-    virtual void add_edge(T vertex_1, T vertex_2, T weight) = 0;//dodanie krawedzi
-    virtual void remove_vertex(T vertex) = 0;//usuwanie wiercholka
-    virtual void remove_edge(T vertex_1, T vertex_2)  = 0;//usuwanie krawedzi
-    virtual std::vector<T> get_neighbors(T vertex)  = 0;//Wyswietlenie wszystkich sasiadow danego wezla 
-    
-    // // TODO: implement all required methods
+public:
+
+    virtual void add_vertex(int vertex) = 0;
+    virtual void add_edge(int vertex_1, int vertex_2, int weight) = 0; // Metoda dodająca krawędź
+    virtual int has_edge(int vertex_1, int vertex_2) const = 0; // Metoda sprawdzająca istnienie krawędzi
+    virtual std::vector<int> getNeighbors(int vertex) const = 0; // Metoda zwracająca sąsiadów wierzchołka
 };
 
 #endif /* GRAPH_HPP_ */
